2dcontext/image-smoothing/2d.imagesmoothing.html

<!DOCTYPE html>
<html>
  <head>
    <title>2dcontext image smoothing test</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <link rel="stylesheet" href="/resources/testharness.css" media="all">
    <link rel="author" title="ShinHyunjin" href="mailto:jini7927@gmail.com">
    <link rel="help" href="http://www.w3.org/html/wg/drafts/2dcontext/html5_canvas/#image-smoothing">
  </head>
  <body>
    <div id="log"></div>
    <canvas id="test_canvas"></canvas>

    <script>    
    (function() {
      test(function() {
        var canvas = document.getElementById('test_canvas');
        var ctx = canvas.getContext('2d');

        //console.log("1:" +  ctx.imageSmoothingEnabled);
        assert_true(ctx.imageSmoothingEnabled);

        test(function() {
        var canvas = document.getElementById('test_canvas');
        var ctx = canvas.getContext('2d');

        ctx.imageSmoothingEnabled = true;
        ctx.imageSmoothingEnabled = false;
        ctx.imageSmoothingEnabled = true;
        //console.log("2:" + ctx.imageSmoothingEnabled);
        assert_true(ctx.imageSmoothingEnabled);
        }, "On getting the imageSmoothingEnabled attribute, the user agent must return the last value it was set to.");

        test(function() {
          var canvas = document.getElementById('test_canvas');
          var ctx = canvas.getContext('2d');

          ctx.imageSmoothingEnabled = false;
          //console.log("3:" + ctx.imageSmoothingEnabled);
          assert_false(ctx.imageSmoothingEnabled); 
        }, "On setting the imageSmoothingEnabled attribute, it must be set to the new value.");
      }, "When the CanvasRenderingContext2D object is created, the attribute must be set to true.");      
    })();
    </script>
  </body>
</html>
